refactoring v2
