+++ internal/api/v2/v2_test.go
@@ -1,4 +1,4 @@
-package api
+package v2
 
 import (
 	"database/sql/driver"
@@ -12,46 +12,78 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"go.uber.org/zap"
 
 	"github.com/stackmon/otc-status-dashboard/internal/api/errors"
 	"github.com/stackmon/otc-status-dashboard/internal/db"
 )
 
-var testAPI *API
-var mock sqlmock.Sqlmock
-
 func TestGetIncidentsHandler(t *testing.T) {
-	initTests(t)
+	r, m := initTests(t)
 
 	str := "2024-09-01T11:45:26.371Z"
 
 	testTime, err := time.Parse(time.RFC3339, str)
 	require.NoError(t, err)
 
-	prepareDB(t, testTime)
+	prepareDB(t, m, testTime)
 
 	var response = `{"data":[{"id":1,"title":"Incident title","impact":0,"components":[150],"start_date":"%s","system":false,"updates":[{"id":1,"status":"resolved","text":"Issue solved.","timestamp":"%s"}]}]}`
 
 	w := httptest.NewRecorder()
-	req, _ := http.NewRequest(http.MethodGet, "/v1/incidents", nil)
-	testAPI.r.ServeHTTP(w, req)
+	req, _ := http.NewRequest(http.MethodGet, "/v2/incidents", nil)
+
+	r.ServeHTTP(w, req)
 
 	assert.Equal(t, 200, w.Code)
-
 	assert.Equal(t, fmt.Sprintf(response, str, str), w.Body.String())
 }
 
 func TestReturn404Handler(t *testing.T) {
-	initTests(t)
+	r, _ := initTests(t)
 	w := httptest.NewRecorder()
 	req, _ := http.NewRequest(http.MethodGet, "/anyendpoint", nil)
-	testAPI.r.ServeHTTP(w, req)
+	r.ServeHTTP(w, req)
 
 	assert.Equal(t, 404, w.Code)
 	assert.Equal(t, `{"errMsg":"page not found"}`, w.Body.String())
 }
 
-func prepareDB(t *testing.T, testTime time.Time) {
+func initTests(t *testing.T) (*gin.Engine, sqlmock.Sqlmock) {
+	t.Helper()
+
+	t.Log("start initialisation")
+	d, m, err := db.NewWithMock()
+	require.NoError(t, err)
+
+	gin.SetMode(gin.TestMode)
+	r := gin.Default()
+	r.NoRoute(errors.Return404)
+
+	log, _ := zap.NewDevelopment()
+	initRoutes(t, r, d, log)
+
+	return r, m
+}
+
+func initRoutes(t *testing.T, c *gin.Engine, dbInst *db.DB, log *zap.Logger) {
+	t.Helper()
+
+	v2Api := c.Group("v2")
+	{
+		v2Api.GET("components", GetComponentsStatusHandler(dbInst, log))
+		v2Api.GET("components/:id", GetComponentHandler(dbInst, log))
+		v2Api.GET("component_status", GetComponentsStatusHandler(dbInst, log))
+		v2Api.POST("component_status", PostComponentStatusHandler(dbInst, log))
+
+		v2Api.GET("incidents", GetIncidentsHandler(dbInst, log))
+		v2Api.POST("incidents", PostIncidentHandler(dbInst, log))
+		v2Api.GET("incidents/:id", GetIncidentHandler(dbInst, log))
+		v2Api.PATCH("incidents/:id", PatchIncidentHandler(dbInst, log))
+	}
+}
+
+func prepareDB(t *testing.T, mock sqlmock.Sqlmock, testTime time.Time) {
 	t.Helper()
 
 	rows := sqlmock.NewRows([]string{"id", "text", "start_date", "end_date", "impact", "system"}).
@@ -86,24 +118,4 @@ func prepareDB(t *testing.T, testTime time.Time) {
 	mock.ExpectQuery("^SELECT (.+) FROM \"component_attribute\"").WillReturnRows(rowsCompAttr)
 
 	mock.NewRowsWithColumnDefinition()
-}
-
-func initTests(t *testing.T) {
-	t.Helper()
-
-	if testAPI != nil && mock != nil {
-		t.Log("testAPI and mock are initialized")
-	}
-
-	t.Log("start initialisation")
-	r := gin.Default()
-	r.Use(ErrorHandle())
-	r.NoRoute(errors.Return404)
-
-	d, m, err := db.NewWithMock()
-	require.NoError(t, err)
-
-	testAPI = &API{r: r, db: d}
-	testAPI.initRoutes()
-	mock = m
 }
